dont puke on ourselves if no man dir exists

+++ make-tarball.sh
@@ -18,13 +18,17 @@ rm -f glibc-${gver}-*.tar.bz2
 mkdir -p tmp/patches
 cp -r ${gver}/*.patch tmp/patches/ || exit 1
 
-mkdir tmp/man
-cp -r ${gver}/man/* tmp/man/ || exit 1
+if [[ -d ${gver}/man ]] ; then
+	mkdir tmp/man
+	cp -r ${gver}/man/* tmp/man/ || exit 1
+fi
 
 tar -jcf glibc-${gver}-patches-${pver}.tar.bz2 \
 	-C tmp patches || exit 1
-tar -jcf glibc-${gver}-manpages.tar.bz2 \
-	-C tmp man || exit 1
+if [[ -d ${gver}/man ]] ; then
+	tar -jcf glibc-${gver}-manpages.tar.bz2 \
+		-C tmp man || exit 1
+fi
 rm -r tmp
 
 du -b *.tar.bz2
